+++ src/main/java/com/ksu/soccerserver/find/FindController.java
@@ -5,6 +5,7 @@ import com.ksu.soccerserver.account.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -14,6 +15,7 @@ import org.springframework.web.server.ResponseStatusException;
 public class FindController {
 
     private final AccountRepository accountRepository;
+    private final PasswordEncoder passwordEncoder;
 
     @GetMapping("/email")
     public ResponseEntity<?> findEmail(
@@ -34,7 +36,14 @@ public class FindController {
                 .orElseThrow(() -> new ResponseStatusException (HttpStatus.NOT_FOUND, "해당하는 사용자 정보를 찾을 수 없습니다."));
 
         return new ResponseEntity<>(foundAccount, HttpStatus.OK);
-        //TODO [ AccountController: @Put /api/accounts/{accountId}/ @RequestBody - {changePw} ]
     }
 
+    // 비밀번호 변경
+    @PutMapping("/changePW")
+    public ResponseEntity<?> changePW(@RequestBody Account account) {
+        Account changeAccount = accountRepository.findByEmail(account.getEmail()).get();
+        changeAccount.changePW(passwordEncoder.encode(account.getPassword()));
+
+        return new ResponseEntity<> (accountRepository.save(changeAccount), HttpStatus.OK);
+    }
 }
